+++ POO/src/module-info.java
@@ -1,0 +1,9 @@
+/**
+ * 
+ */
+/**
+ * 
+ */
+module POO {
+	requires java.desktop;
+}
