--- README.md
@@ -1,5 +1,0 @@
-# Proyecto-POO
-App de química 
-
-
-Keyner por favor usa variables descriptivas
